+++ sdk/src/main/java/com/hedera/hashgraph/sdk/AccountInfo.java
@@ -120,13 +120,15 @@ public final class AccountInfo {
 
     public final int maxAutomaticTokenAssociations;
 
+    @Nullable
+    public final PublicKey aliasKey;
+
     /**
      * The ledger ID the response was returned from; please see <a href="https://github.com/hashgraph/hedera-improvement-proposal/blob/master/HIP/hip-198.md">HIP-198</a> for the network-specific IDs.
      */
     public final LedgerId ledgerId;
 
-    @Nullable
-    public final PublicKey aliasKey;
+    public long ethereumNonce;
 
     @Deprecated
     public final List<HbarAllowance> hbarAllowances;
@@ -156,7 +158,8 @@ public final class AccountInfo {
         long ownedNfts,
         int maxAutomaticTokenAssociations,
         @Nullable PublicKey aliasKey,
-        LedgerId ledgerId
+        LedgerId ledgerId,
+        long ethereumNonce
     ) {
         this.accountId = accountId;
         this.contractAccountId = contractAccountId;
@@ -177,6 +180,7 @@ public final class AccountInfo {
         this.maxAutomaticTokenAssociations = maxAutomaticTokenAssociations;
         this.aliasKey = aliasKey;
         this.ledgerId = ledgerId;
+        this.ethereumNonce = ethereumNonce;
         this.hbarAllowances = Collections.emptyList();
         this.tokenAllowances = Collections.emptyList();
         this.tokenNftAllowances = Collections.emptyList();
@@ -222,7 +226,8 @@ public final class AccountInfo {
             accountInfo.getOwnedNfts(),
             accountInfo.getMaxAutomaticTokenAssociations(),
             aliasKey,
-            LedgerId.fromByteString(accountInfo.getLedgerId())
+            LedgerId.fromByteString(accountInfo.getLedgerId()),
+            accountInfo.getEthereumNonce()
         );
     }
 
@@ -250,7 +255,8 @@ public final class AccountInfo {
             .setMemo(accountMemo)
             .setOwnedNfts(ownedNfts)
             .setMaxAutomaticTokenAssociations(maxAutomaticTokenAssociations)
-            .setLedgerId(ledgerId.toByteString());
+            .setLedgerId(ledgerId.toByteString())
+            .setEthereumNonce(ethereumNonce);
 
         if (contractAccountId != null) {
             accountInfoBuilder.setContractAccountID(contractAccountId);
@@ -289,6 +295,7 @@ public final class AccountInfo {
             .add("maxAutomaticTokenAssociations", maxAutomaticTokenAssociations)
             .add("aliasKey", aliasKey)
             .add("ledgerId", ledgerId)
+            .add("ethereumNonce", ethereumNonce)
             .toString();
     }