+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * `ContractUpdateTransaction.[get|set|clear]AutoRenewAccountId()`
  * `ContractCreateFlow.[get|set]MaxAutomaticTokenAssociations()`
  * `ContractCreateFlow.[get|set]AutoRenewAccountId()`
+ * `AccountInfo.ethereumNonce`
+ * `ContractCallResult.senderAccountId`
+ * `ContractCallQuery.[get|set]SenderAccountId()`
+ * `TransactionRecord.ethereumHash`
+ * `EthereumTransaction`
 
 ## v2.12.0